+++ cmd/zetae2e/local/evm.go
@@ -21,6 +21,7 @@ func startEVMTests(eg *errgroup.Group, conf config.Config, deployerRunner *runne
 		e2etests.TestETHDepositAndCallBigPayloadName,
 		e2etests.TestETHDepositFastConfirmationName,
 		e2etests.TestETHWithdrawName,
+		e2etests.TestETHWithdrawCustomGasLimitName,
 		e2etests.TestETHWithdrawAndArbitraryCallName,
 		e2etests.TestETHWithdrawAndCallName,
 		e2etests.TestETHWithdrawAndCallBigPayloadName,
@@ -94,6 +95,7 @@ func startEVMTests(eg *errgroup.Group, conf config.Config, deployerRunner *runne
 		),
 	)
 
+	// TODO evm: tmp comment out because of setup difference between zetae2e-ante and zetae2e
 	eg.Go(
 		evmTestRoutine(conf, "zeta", conf.AdditionalAccounts.UserZeta, color.FgHiBlue, deployerRunner, verbose,
 			e2etests.TestZetaDepositName,